requirements.txt
django==3.1.3
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2==2.8.6
pyunicore==0.8.2
